tests/test_requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
docutils==0.16
idna==2.10
ImageHash==4.2.0
imagesize==1.2.0
importlib-metadata==3.4.0
iniconfig==1.1.1
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.5
packaging==20.8
Pillow==8.1.0
pluggy==0.13.1
py==1.10.0
Pygments==2.7.4
pyparsing==2.4.7
pytest==6.2.1
pytz==2020.5
PyWavelets==1.1.1
requests==2.25.1
scipy==1.6.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
zipp==3.4.0